Created new file for main program
